requirements.txt
asgiref==3.8.1
boto3==1.35.87
botocore==1.35.87
certifi==2024.12.14
charset-normalizer==3.4.1
cramjam==2.9.1
distlib==0.4.0
Django==5.1.7
djangorestframework==3.16.0
fastparquet==2024.11.0
filelock==3.18.0
fsspec==2024.12.0
idna==3.10
jmespath==1.0.1
numpy==2.2.1
packaging==24.2
pandas==2.2.3
platformdirs==4.3.8
psycopg2-binary==2.9.10
pyarrow==18.1.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.1.1
pytz==2024.2
requests==2.32.3
requests-aws4auth==1.3.1
s3transfer==0.10.4
setuptools==75.8.0
six==1.17.0
sqlparse==0.5.3
tzdata==2024.2
urllib3==2.3.0
virtualenv==20.32.0
wheel==0.45.1
